refactor to separate internal packages
